requirements.txt
py-cord>=2.7.0rc1
python-dotenv>=1.0.0
wavelink>=3.0.0
aiohttp>=3.9.0
requests>=2.31.0
